requirements/requirements.txt
arrow==1.2.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Django==3.2
django-rest-knox==4.2.0
djangorestframework==3.12.0
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
influxable==1.4.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pathlib==1.0.1
psycopg2==2.9.3
pycparser==2.21
pyhumps==1.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
